pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "qmof-hull"
version = "1.0.0"
description = "Tiny helper package for energy-above-hull phase diagram construction"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-3" }
authors = [
  { name = "Rosen Group" }
]

dependencies = [
  "ase>=3.26.0",
  "fairchem-core>=2.2.0",
  "monty>=2026.2.18",
  "pymatgen>=2024.0.0",
  "pandas>=1.5",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "ruff>=0.0.285"]

[tool.setuptools.package-data]
qmof_thermo = ["py.typed", "*.json"]

[tool.pyright]
include = ["qmof_thermo"]
exclude = ["**/__pycache__"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-p no:warnings", "--import-mode=importlib"]
xfail_strict = true
log_cli_level = "warn"
pythonpath = "src"
testpaths = ["tests"]

[tool.black]
exclude = '''
/(
    \.git
  | \.tox
)/
'''
skip-magic-trailing-comma = true

[tool.isort]
profile = 'black'
skip_gitignore = true

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "except ImportError",
]

[tool.ruff]
lint.select = [
  "A",      # flake8-builtins
  "ARG",    # flake8-unused-arguments
  "ASYNC",  # flake8-async
  "ASYNC1", # flake8-async1
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "DTZ",    # flake8-datetimez
  "E",      # pycodestyle error
  "EXE",    # flake8-executable
  "F",      # pyflakes
  "FA",     # flake8-future-annotations
  "FLY",    # flynt
  "I",      # isort
  "ICN",    # flake8-import-conventions
  "INT",    # flake8-gettext
  "ISC",    # flake8-implicit-str-concat
  "LOG",    # flake8-logging
  "NPY",    # numpy-specific rules
  "PD",     # pandas-vet
  "PERF",   # perflint
  "PIE",    # flake8-pie
  "PL",     # pylint
  "PT",     # flake8-pytest-style
  "PTH",    # flake8-use-pathlib
  "PYI",    # flake8-pyi
  "Q",      # flake8-quotes
  "RET",    # flake8-return
  "RSE",    # flake8-raise
  "RUF",    # Ruff-specific rules
  "SIM",    # flake8-simplify
  "SLOT",   # flake8-slots
  "T20",    # flake8-print
  "TCH",    # flake8-type-checking
  "TID",    # flake8-tidy-imports
  "UP",     # pyupgrade
  "W",      # pycodestyle warning
  "YTT",    # flake8-2020
]
lint.ignore = [
  "E501",   # Line too long
  "ISC001",   # single-line-implicit-string-concatenation
  "PLR",    # Design related pylint codes
  "PERF203",  # try-except-in-loop
  "RET505", # Unnecessary `elif` after `return`
  "PLC0415", # `import` should be at the top-level of a file
  "T201", # print
  "EXE002", #shebang
  "RUF001", # greek letters
  "RUF003", # hyphens
]
src = ["src"]
lint.unfixable = [
  "T20",  # Removes print statements
  "F841", # Removes unused variables
]
lint.pydocstyle.convention = "numpy"
lint.isort.known-first-party = ["qmof_thermo"]
lint.isort.required-imports = ["from __future__ import annotations"]
lint.isort.split-on-trailing-comma=false
extend-include = ["*.ipynb"]

[tool.ruff.format]
docstring-code-format = true
skip-magic-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ANN", "ARG", "D", "E402", "PTH", "S101"]

[tool.docformatter]
pre-summary-newline = true
black = true

[tool.mypy]
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true
no_implicit_optional = false
disable_error_code = "annotate"
